+++ CHANGELOG.md
@@ -9,6 +9,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - edit validation of db names to give the possibility to use name with underscore
 - remove transaction for catalog_product_flat reindexAll
 - add config to toggle the transaction for catalog_product_flat
-- add configurable cron for reindex
+- add configurable cron for reindex of a list of processes
 - add config to toggle the transaction for catalog_product_price